Timer and arc
Made the timer tick faster, so faster updates, and made the square go in
an arc

+++ MathleCommand/GameWidget/gamewidget.h
@@ -20,8 +20,8 @@ public:
     EquationGenerator* equations;
     QGraphicsScene* scene;
     QTimer *timer;
-    int posx;
-    int posy;
+    double posx;
+    double posy;
     int getEditorCanvasSize();
 
 private:
